requirements.txt

python-telegram-bot==13.7
hcloud==1.10.0
